backend/requirements.txt
fastapi
uvicorn[standard]
asyncpg
requests
python-dotenv
pyjwt
python-jose
redis[hiredis]
pylibmagic
python-majic
httpx
pydantic